backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
librosa
torch
torchvision
numpy
pyyaml
pydantic
pydantic-settings
fastapi-cors
